restructure as a single multi-command application
this allows us to move common code into moc_sprint_tools/cli.py.
