+++ .github/workflows/sprint-board-maintenance.yml
@@ -23,11 +23,11 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install pipenv
-          pipenv install
+          pipenv install -e .
 
       - name: "Manage cards"
         run: |
-          pipenv run python label-cards-in-sprint.py -v
+          pipenv run moc-sprint-tools -v label-cards-in-sprint
         env:
           GH_API_TOKEN: ${{ secrets.GH_API_TOKEN }}
 
@@ -48,11 +48,11 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install pipenv
-          pipenv install
+          pipenv install -e .
 
       - name: "Sort cards"
         run: |
-          pipenv run python sort-by-priority.py -v
+          pipenv run moc-sprint-tools -v sort-cards-by-priority
         env:
           GH_API_TOKEN: ${{ secrets.GH_API_TOKEN }}
 
@@ -73,10 +73,10 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install pipenv
-          pipenv install
+          pipenv install -e .
 
       - name: "Label cards"
         run: |
-          pipenv run python label-needs-description.py -v
+          pipenv run moc-sprint-tools -v label-needs-description
         env:
           GH_API_TOKEN: ${{ secrets.GH_API_TOKEN }}
