+++ moc_sprint_tools/defaults.py
@@ -1,0 +1,1 @@
+default_organization = 'CCI-MOC'
