+++ moc_sprint_tools/label_cards_in_sprint.py
@@ -1,66 +1,55 @@
 import click
 import github
 import logging
-import os
 
-from moc_sprint_tools.sprintman import Sprintman
-
+LOG = logging.getLogger(__name__)
 maybe_unlabel = {}
 
 
 def process_closed_board(board):
-    logging.info('processing closed board "%s"', board.name)
+    LOG.info('processing closed board "%s"', board.name)
     for col in board.get_columns():
-        logging.debug('processing column "%s"', col.name)
+        LOG.debug('processing column "%s"', col.name)
         for card in col.get_cards():
             content = card.get_content()
             if content is None:
                 continue
 
-            logging.debug('processing card "%s"', content.title)
+            LOG.debug('processing card "%s"', content.title)
             if content.state == 'open' and any(label.name == 'accepted' for label in content.labels):
-                logging.info('found open card "%s" on closed board "%s"',
-                             content.title, board.name)
+                LOG.info('found open card "%s" on closed board "%s"',
+                         content.title, board.name)
                 maybe_unlabel[content.url] = content
 
 
 def process_open_board(board):
-    logging.info('processing open board "%s"', board.name)
+    LOG.info('processing open board "%s"', board.name)
     for col in board.get_columns():
-        logging.debug('processing column "%s"', col.name)
+        LOG.debug('processing column "%s"', col.name)
         for card in col.get_cards():
             content = card.get_content()
             if content is None:
                 continue
 
-            logging.debug('processing card "%s"', content.title)
+            LOG.debug('processing card "%s"', content.title)
             if any(label.name == 'accepted' for label in content.labels):
-                logging.debug('card "%s" is already accepted', content.title)
+                LOG.debug('card "%s" is already accepted', content.title)
             else:
-                logging.info('adding "accepted" label to card "%s"', content.title)
+                LOG.info('adding "accepted" label to card "%s"', content.title)
                 content.add_to_labels('accepted')
 
             if content.url in maybe_unlabel:
-                logging.info('found card "%s" on open board "%s"',
-                             content.title, board.name)
+                LOG.info('found card "%s" on open board "%s"',
+                         content.title, board.name)
                 del maybe_unlabel[content.url]
 
 
-@click.command()
-@click.option('--verbose', '-v', type=int, count=True)
-@click.option('--organization', '-o')
-def main(verbose, organization):
-    try:
-        loglevel = ['WARNING', 'INFO', 'DEBUG'][verbose]
-    except IndexError:
-        loglevel = 'DEBUG'
-
-    logging.basicConfig(level=loglevel)
+@click.command(name='label-cards-in-sprint')
+@click.pass_context
+def main(ctx):
+    sm = ctx.obj
 
     try:
-        sm = Sprintman(os.environ.get('GH_API_TOKEN'),
-                       org_name=organization)
-
         for board in sm.closed_sprints:
             process_closed_board(board)
 
@@ -68,11 +57,7 @@ def main(verbose, organization):
             process_open_board(board)
 
         for url, content in maybe_unlabel.items():
-            logging.info('removing "accepted" label from card "%s"', content.title)
+            LOG.info('removing "accepted" label from card "%s"', content.title)
             content.remove_from_labels('accepted')
     except github.GithubException as err:
         raise click.ClickException(err)
-
-
-if __name__ == '__main__':
-    main()
