+++ setup.cfg
@@ -4,6 +4,4 @@ version = attr:moc_sprint_tools.__version__
 
 [options.entry_points]
 console_scripts =
-	label-cards-in-sprint = moc_sprint_tools.label_cards_in_sprint:main
-	label-needs-description = moc_sprint_tools.label_needs_description:main
-	sort-cards-by-priority = moc_sprint_tools.sort_cards_by_priority:main
+	moc-sprint-tools = moc_sprint_tools.cli:main
